requirements.txt
celery==5.4.0
Django==5.0.7
django-json-widget==2.0.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.28.0
openai==1.54.4
pydantic==2.9.2
requests==2.31.0
django-debug-toolbar==4.4.6
django-filter==24.3

